Revert "Refactor directory and complete readme"
