+++ client/src/Authentication/Signin.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { reduxForm } from 'redux-form';
-import * as actions from './authAction';
+import * as actions from './AuthActions';
 import axios from 'axios';
 
 class Signin extends Component {
