+++ client/src/Authentication/Signout.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import * as actions from './authAction';
-import Header from '../App/header';
+import * as actions from './AuthActions';
+import Header from '../App/Header';
 
 class Signout extends Component {
   componentWillMount() {
