+++ client/src/Dashboard/Dashboard.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import Header from '../App/header';
-import Quiz from '../Quiz/quiz.js';
-import Meter from '../Meter/meter';
+import * as actions from '../Authentication/AuthActions';
+import Header from '../App/Header';
+import Quiz from '../Quiz';
+import Meter from '../Meter';
 
 export default class Dashboard extends Component {
