+++ client/src/Quiz/index.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { reduxForm } from 'redux-form';
-import * as actions from './quizAction';
-import './quiz.css';
+import * as actions from './QuizActions';
+import './index.css';
 
 class Quiz extends Component {
   constructor(props) {
