+++ client/src/rootReducer/index.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 import { reducer as form } from 'redux-form';
-import authReducer from '../../components/Authentication/authReducer';
-import quizReducer from '../../components/Quiz/quizReducer';
-import meterReducer from '../../components/Meter/meterReducer';
+import authReducer from '../Authentication/authReducer';
+import quizReducer from '../Quiz/quizReducer';
+import meterReducer from '../Meter/meterReducer';
 
 const rootReducer = combineReducers({
   form,
